requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic>=2.7.0,<3.0.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Computer Vision
ultralytics>=8.3.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy<2.0.0,>=1.24.3

# Data processing
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
requests>=2.32.5,<3.0.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1

# Hugging Face Datasets (for Kanops dataset)
datasets==2.14.6
huggingface-hub==0.17.3

# Hub SDK for Ultralytics Hub integration
hub-sdk>=0.0.12

